# packages/api-client/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "nexaOSweb Brain",
    "version": "0.1.0"
  },
  "paths": {
...
    "/projects/{project_id}": {
      "patch": {
        "tags": [
          "projects"
        ],
        "summary": "Rename Project",
        "operationId": "rename_project_projects__project_id__patch",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ProjectRenameRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "projects"
        ],
        "summary": "Delete Project",
        "operationId": "delete_project_projects__project_id__delete",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/projects/{project_id}/duplicate": {
      "post": {
        "tags": [
          "projects"
        ],
        "summary": "Duplicate Project",
        "operationId": "duplicate_project_projects__project_id__duplicate_post",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Project Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
    "/research/projects": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "List Research Projects",
        "operationId": "list_research_projects_research_projects_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/ResearchProjectRead"
                  },
                  "type": "array",
                  "title": "Response List Research Projects Research Projects Get"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Create Research Project",
        "operationId": "create_research_project_research_projects_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchProjectCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchProjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research/projects/{research_id}": {
      "patch": {
        "tags": [
          "research"
        ],
        "summary": "Update Research Project",
        "operationId": "update_research_project_research_projects__research_id__patch",
        "parameters": [
          {
            "name": "research_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Research Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResearchProjectUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchProjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "research"
        ],
        "summary": "Delete Research Project",
        "operationId": "delete_research_project_research_projects__research_id__delete",
        "parameters": [
          {
            "name": "research_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Research Id"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research/projects/{research_id}/duplicate": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Duplicate Research Project",
        "operationId": "duplicate_research_project_research_projects__research_id__duplicate_post",
        "parameters": [
          {
            "name": "research_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Research Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchProjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research/generate-config": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Generate Research Config",
        "operationId": "generate_research_config_research_generate_config_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/GenerateConfigRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GenerateConfigResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research/{research_id}/attach": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Attach",
        "operationId": "attach_research__research_id__attach_post",
        "parameters": [
          {
            "name": "research_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Research Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AttachRequest"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research/{research_id}/detach": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Detach",
        "operationId": "detach_research__research_id__detach_post",
        "parameters": [
          {
            "name": "research_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Research Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research/{research_id}/runs": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Trigger Run",
        "operationId": "trigger_run_research__research_id__runs_post",
        "parameters": [
          {
            "name": "research_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Research Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ResearchRunRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      },
      "get": {
        "tags": [
          "research"
        ],
        "summary": "List Runs",
        "operationId": "list_runs_research__research_id__runs_get",
        "parameters": [
          {
            "name": "research_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Research Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResearchRunRead"
                  },
                  "title": "Response List Runs Research  Research Id  Runs Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research/{research_id}/findings": {
      "get": {
        "tags": [
          "research"
        ],
        "summary": "List Findings",
        "operationId": "list_findings_research__research_id__findings_get",
        "parameters": [
          {
            "name": "research_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Research Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ResearchFindingRead"
                  },
                  "title": "Response List Findings Research  Research Id  Findings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research/findings/{finding_id}/to-task": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Finding To Task",
        "operationId": "finding_to_task_research_findings__finding_id__to_task_post",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Finding Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research/findings/{finding_id}/to-update": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Finding To Update",
        "operationId": "finding_to_update_research_findings__finding_id__to_update_post",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Finding Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProjectUpdateRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/research/findings/{finding_id}/to-knowledge": {
      "post": {
        "tags": [
          "research"
        ],
        "summary": "Finding To Knowledge",
        "operationId": "finding_to_knowledge_research_findings__finding_id__to_knowledge_post",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Finding Id"
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/KnowledgeEntryRead"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ProjectRead": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "item_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Item Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "slug": {
            "type": "string",
            "title": "Slug"
          },
          "stage": {
            "type": "string",
            "title": "Stage"
          },
          "mode": {
            "type": "string",
            "title": "Mode"
          },
          "plan_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Plan Path"
          },
          "plan_json": {
            "additionalProperties": true,
            "type": "object",
            "title": "Plan Json"
          },
          "build_destination": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Build Destination"
          },
          "selected_integrations": {
            "items": {},
            "type": "array",
            "title": "Selected Integrations"
          },
          "workspace": {
            "additionalProperties": true,
            "type": "object",
            "title": "Workspace"
          },
          "research_target_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Research Target Id"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "format": "date-time",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "item_id",
          "name",
          "slug",
          "stage",
          "mode",
          "plan_path",
          "plan_json",
          "build_destination",
          "selected_integrations",
          "workspace",
          "research_target_id",
          "created_at",
          "updated_at"
        ],
        "title": "ProjectRead"
      },
      "ProjectRenameRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          }
        },
        "type": "object",
        "required": [
          "name"
        ],
        "title": "ProjectRenameRequest"
      },
...
    }
  }
}